Move the Ollama and OpenAI wire types into serve/types.rs
